Exported the missing SQLite package code
