+++ src/packages/emmett-sqlite/src/connection/index.ts
@@ -1,0 +1,1 @@
+export * from './sqliteConnection';
