+++ src/packages/emmett-sqlite/src/connection/sqliteConnection.ts
