+++ src/packages/emmett-sqlite/src/eventStore/SQLiteEventStore.e2e.spec.ts
@@ -10,7 +10,7 @@ import { afterEach, beforeEach, describe, it } from 'node:test';
 import path from 'path';
 import { fileURLToPath } from 'url';
 import { v4 as uuid } from 'uuid';
-import { InMemorySQLiteDatabase, sqliteConnection } from '../sqliteConnection';
+import { InMemorySQLiteDatabase, sqliteConnection } from '../connection';
 import {
   type DiscountApplied,
   type PricedProductItem,
