+++ src/packages/emmett-sqlite/src/eventStore/index.ts
@@ -1,0 +1,2 @@
+export * from './schema';
+export * from './SQLiteEventStore';
