+++ src/packages/emmett-sqlite/src/eventStore/schema/index.ts
@@ -1,14 +1,4 @@
-import { type SQLiteConnection } from '../../sqliteConnection';
-import { messagesTableSQL, streamsTableSQL } from './tables';
-
+export * from './appendToStream';
+export * from './readStream';
 export * from './tables';
-
-export const schemaSQL: string[] = [streamsTableSQL, messagesTableSQL];
-
-export const createEventStoreSchema = async (
-  db: SQLiteConnection,
-): Promise<void> => {
-  for (const sql of schemaSQL) {
-    await db.command(sql);
-  }
-};
+export * from './typing';
