+++ src/packages/emmett-sqlite/src/eventStore/schema/readStream.int.spec.ts
@@ -12,7 +12,7 @@ import {
   InMemorySQLiteDatabase,
   sqliteConnection,
   type SQLiteConnection,
-} from '../../sqliteConnection';
+} from '../../connection';
 import { appendToStream } from './appendToStream';
 import { readStream } from './readStream';
