+++ src/packages/emmett-sqlite/src/index.ts
@@ -1,0 +1,2 @@
+export * from './connection';
+export * from './eventStore';
